Split solidgauge series into multiple files
