+++ ts/Series/SolidGauge/SolidGaugePointOptions.d.ts
@@ -1,0 +1,36 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type GaugePointOptions from '../Gauge/GaugePointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface SolidGaugePointOptions extends GaugePointOptions {
+    innerRadius?: (number | string);
+    radius?: (number | string);
+}
+
+/* *
+ *
+ *  Default export
+ *
+ * */
+
+export default SolidGaugePointOptions;
